# docs/swagger.yaml
definitions:
  type_api.BaseResponse-any:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  type_api.BaseResponse-type_api_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/type_api.LoginResponse'
      message:
        type: string
      status:
        type: integer
    type: object
  type_api.BaseResponseError:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  type_api.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  type_api.LoginResponse:
    properties:
      role:
        description: نقش کاربر، مثلاً "admin" یا "reseller"
        items:
          type: string
        type: array
      token:
        type: string
    required:
    - role
    - token
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: User login with username and password
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/type_api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/type_api.BaseResponse-type_api_LoginResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/type_api.BaseResponseError'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/type_api.BaseResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/type_api.BaseResponseError'
      summary: User Login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Revoke the current user's session token
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/type_api.BaseResponse-any'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/type_api.BaseResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/type_api.BaseResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/type_api.BaseResponseError'
      summary: User Logout
      tags:
      - Auth
swagger: "2.0"